fix(visual-editing): resolve inline schema types when getting fields
