+++ packages/visual-editing/src/ui/ElementOverlay.tsx
@@ -27,7 +27,6 @@ import type {
   SanityStegaNode,
 } from '../types'
 import {usePreviewSnapshots} from './preview/usePreviewSnapshots'
-import {getField, getSchemaType} from './schema/schema'
 import {useSchema} from './schema/useSchema'
 
 export interface ElementOverlayProps {
@@ -182,9 +181,9 @@ const ComponentWrapper: FunctionComponent<{
 const ElementOverlayInner: FunctionComponent<ElementOverlayProps> = (props) => {
   const {element, focused, componentResolver, node, showActions, draggable} = props
 
-  const {schema, resolvedTypes} = useSchema()
-  const schemaType = getSchemaType(node, schema)
-  const {field, parent} = getField(node, schemaType, resolvedTypes)
+  const {getField, getType} = useSchema()
+  const schemaType = getType(node)
+  const {field, parent} = getField(node)
 
   const href = 'path' in node ? createIntentLink(node) : node.href
